migrated all delegate providers to serializers
