--- Manatee.Json/Serialization/Internal/AutoRegistration/NullableSerializationDelegateProvider.cs
@@ -1,32 +1,0 @@
-using System;
-using System.Reflection;
-
-namespace Manatee.Json.Serialization.Internal.AutoRegistration
-{
-	internal class NullableSerializationDelegateProvider : SerializationDelegateProviderBase
-	{
-		public override bool CanHandle(Type type)
-		{
-			return type.GetTypeInfo().IsGenericType && type.GetGenericTypeDefinition() == typeof(Nullable<>);
-		}
-
-		private static JsonValue _Encode<T>(T? nullable, JsonSerializer serializer)
-			where T : struct
-		{
-			if (!nullable.HasValue) return JsonValue.Null;
-			var encodeDefaultValues = serializer.Options.EncodeDefaultValues;
-			serializer.Options.EncodeDefaultValues = Equals(nullable.Value, default (T));
-			var json = serializer.Serialize(nullable.Value);
-			serializer.Options.EncodeDefaultValues = encodeDefaultValues;
-			return json;
-		}
-		private static T? _Decode<T>(JsonValue json, JsonSerializer serializer)
-			where T : struct
-		{
-			if (json == JsonValue.Null)
-				return null;
-			T? nullable = serializer.Deserialize<T>(json);
-			return nullable;
-		}
-	}
-}
