+++ Manatee.Json/Serialization/Internal/IPrioritizedSerializer.cs
@@ -1,0 +1,7 @@
+﻿namespace Manatee.Json.Serialization.Internal
+{
+	internal interface IPrioritizedSerializer : ISerializer
+	{
+		int Priority { get; }
+	}
+}
