+++ Manatee.Json/Serialization/Internal/Serializers/BooleanSerializer.cs
@@ -2,11 +2,13 @@
 
 namespace Manatee.Json.Serialization.Internal.Serializers
 {
-	internal class BooleanSerializer : ISerializer
+	internal class BooleanSerializer : IPrioritizedSerializer
 	{
+		public int Priority => int.MinValue;
+
 		public bool ShouldMaintainReferences => false;
 
-		public bool Handles(Type type, JsonSerializerOptions options)
+		public bool Handles(Type type, JsonSerializerOptions options, JsonValue json)
 		{
 			return type == typeof(bool);
 		}
