+++ Manatee.Json/Serialization/Internal/Serializers/ReferencingSerializer.cs
@@ -13,7 +13,7 @@ namespace Manatee.Json.Serialization.Internal.Serializers
 			_innerSerializer = innerSerializer;
 		}
 
-		public bool Handles(Type type, JsonSerializerOptions options)
+		public bool Handles(Type type, JsonSerializerOptions options, JsonValue json)
 		{
 			return true;
 		}
