+++ Manatee.Json/Serialization/SerializerFactory.cs
@@ -20,12 +20,19 @@ namespace Manatee.Json.Serialization
 		private static readonly StringSerializer _stringSerializer;
 		private static readonly Dictionary<Type, ISerializer> _library;
 		private static readonly SchemaSerializer _schemaSerializer;
-		private static readonly UriSerializer _uriSerializer;
-
-		private static List<ISerializer> UserSerializers { get; }
+		private static readonly List<ISerializer> _userSerializers;
+		private static readonly List<ISerializer> _serializers;
+		private static readonly TypeInfo[] _dependentSerializers =
+			{
+				typeof(ReferencingSerializer).GetTypeInfo(),
+				typeof(SchemaValidator).GetTypeInfo(),
+				typeof(DefaultValueSerializer).GetTypeInfo()
+			};
+		private static readonly IEnumerable<ISerializer> _orderedSerializers;
 
 		static SerializerFactory()
 		{
+			// todo: reduce this to a single, prioritized list
 			_autoSerializer = new AutoSerializer();
 			_booleanSerializer = new BooleanSerializer();
 			_enumNameSerializer = new EnumNameSerializer();
@@ -35,7 +42,6 @@ namespace Manatee.Json.Serialization
 			_registeredObjectSerializer = new RegisteredObjectSerializer();
 			_stringSerializer = new StringSerializer();
 			_schemaSerializer = new SchemaSerializer();
-			_uriSerializer = new UriSerializer();
 			_library = new Dictionary<Type, ISerializer>
 				{
 					{typeof (sbyte), _numericSerializer},
@@ -51,47 +57,63 @@ namespace Manatee.Json.Serialization
 					{typeof (double), _numericSerializer},
 					{typeof (decimal), _numericSerializer},
 					{typeof (bool), _booleanSerializer},
-					{typeof (string), _stringSerializer},
-					{typeof (Uri), _uriSerializer }
+					{typeof (string), _stringSerializer}
 				};
-			UserSerializers = new List<ISerializer>();
+			_userSerializers = new List<ISerializer>
+				{
+					new ArraySerializer(),
+				};
+			_serializers = typeof(ISerializer).GetTypeInfo()
+			                                     .Assembly
+			                                     .DefinedTypes
+			                                     .Where(t => t.ImplementedInterfaces.Contains(typeof(ISerializer)) &&
+			                                                 t.IsClass && !t.IsAbstract)
+			                                     .Except(_dependentSerializers)
+			                                     .Select(t => Activator.CreateInstance(t.AsType()))
+			                                     .Cast<ISerializer>()
+			                                     .ToList();
+			_orderedSerializers = _serializers.OrderBy(s => (s as IPrioritizedSerializer)?.Priority ?? 0);
+		}
+
+		public static void AddSerializer(ISerializer serializer)
+		{
+			var existing = _serializers.FirstOrDefault(s => s.GetType() == serializer.GetType());
+			if (existing == null)
+				_serializers.Add(serializer);
+		}
+		public static void RemoveSerializer<T>()
+			where T : ISerializer
+		{
+			var serializer = _serializers.OfType<T>().FirstOrDefault();
+			_serializers.Remove(serializer);
 		}
 
 		internal static ISerializer GetSerializer(Type typeToSerialize, JsonSerializer serializer, JsonValue json = null)
 		{
 			typeToSerialize = serializer.AbstractionMap.GetMap(typeToSerialize);
-			var typeInfo = typeToSerialize.GetTypeInfo();
-			if (typeof (IJsonSchema).GetTypeInfo().IsAssignableFrom(typeInfo))
-				return _BuildSerializer(_schemaSerializer, typeInfo);
-			var byUser = UserSerializers.FirstOrDefault(s => s.Handles(typeToSerialize, serializer.Options));
-			if (byUser != null)
-				return _BuildSerializer(byUser, typeInfo);
-			if (serializer.CustomSerializations.IsRegistered(typeToSerialize))
-				return _BuildSerializer(_registeredObjectSerializer, typeInfo);
-			if (typeof (IJsonSerializable).GetTypeInfo().IsAssignableFrom(typeInfo))
-				return _BuildSerializer(_jsonSerializableSerializer, typeInfo);
-			if (typeof (Enum).GetTypeInfo().IsAssignableFrom(typeInfo))
-			{
-				if (json != null)
-				{
-					if (json.Type == JsonValueType.Number)
-						return _BuildSerializer(_enumValueSerializer, typeInfo);
-					if (json.Type == JsonValueType.String)
-						return _BuildSerializer(_enumNameSerializer, typeInfo);
-				}
-				switch (serializer.Options.EnumSerializationFormat)
-				{
-					case EnumSerializationFormat.AsInteger:
-						return _BuildSerializer(_enumValueSerializer, typeInfo);
-					case EnumSerializationFormat.AsName:
-						return _BuildSerializer(_enumNameSerializer, typeInfo);
-					default:
-						throw new ArgumentOutOfRangeException();
-				}
-			}
-			return _BuildSerializer(_library.TryGetValue(typeToSerialize, out var internalSerializer)
-				                        ? internalSerializer
-										: _autoSerializer, typeInfo);
+			var theChosenOne = _orderedSerializers.FirstOrDefault(s => s.Handles(typeToSerialize, serializer.Options, json));
+
+			return _BuildSerializer(theChosenOne, typeToSerialize.GetTypeInfo());
+
+			//if (typeof(Enum).GetTypeInfo().IsAssignableFrom(typeInfo))
+			//{
+			//	if (json != null)
+			//	{
+			//		if (json.Type == JsonValueType.Number)
+			//			return _BuildSerializer(_enumValueSerializer, typeInfo);
+			//		if (json.Type == JsonValueType.String)
+			//			return _BuildSerializer(_enumNameSerializer, typeInfo);
+			//	}
+			//	switch (serializer.Options.EnumSerializationFormat)
+			//	{
+			//		case EnumSerializationFormat.AsInteger:
+			//			return _BuildSerializer(_enumValueSerializer, typeInfo);
+			//		case EnumSerializationFormat.AsName:
+			//			return _BuildSerializer(_enumNameSerializer, typeInfo);
+			//		default:
+			//			throw new ArgumentOutOfRangeException();
+			//	}
+			//}
 		}
 		internal static ITypeSerializer GetTypeSerializer()
 		{
@@ -103,8 +125,7 @@ namespace Manatee.Json.Serialization
 			if (!typeInfo.IsValueType && innerSerializer.ShouldMaintainReferences)
 				innerSerializer = new ReferencingSerializer(innerSerializer);
 
-			return new SchemaValidator(
-				new DefaultValueSerializer(innerSerializer));
+			return new SchemaValidator(new DefaultValueSerializer(innerSerializer));
 		}
 	}
 }
